oneplus6: drop in op6 files
https://github.com/android-linux-stable/op6/commit/e2d5e6cd35eb711a1ae848516c7bff204e2f9a9a#diff-131bfba24104ddf2dd08e64453e55292a86c7416a5df859f7e0937ab698696d7

+++ drivers/media/platform/msm/camera/cam_jpeg/jpeg_hw/cam_jpeg_hw_mgr.h
@@ -44,7 +44,7 @@
 struct cam_jpeg_process_frame_work_data_t {
 	uint32_t type;
 	void *data;
-	uintptr_t request_id;
+	uint64_t request_id;
 };
 
 /**
